feat:프로필 사진 초기화 구현
이미지 취소, 입력 폼 제출 시 초기화
